Java library (but not yet GUI) work with predefined list of values for String parameter
The Transformer (that convert object structure to XML file) now work correctly (it print indentation, now).
Parameter class now reuse the Element builder instead of XMLString builder.
It allow subclass of Parameter to override method toXMLElement() that is easier to implement than overriding the method toXMLString().
The methods toXMLString() in Parameter and ParameterFile are marked as @Deprecated to avoid using them, but these methods are not removed.